chore: final cleanup for v1 vanilla version
