+++ src/features/reminder/domain/ReminderStore.ts
@@ -1,6 +1,6 @@
 import { Store } from '@core/Store';
 import { ReminderSectionData } from './reminder';
-import { STORAGE_KEYS, CATEGORIES } from '@src/shared/constants';
+import { STORAGE_KEYS, Category } from '@src/shared/constants';
 
 interface ReminderState {
   sections: ReminderSectionData[];
@@ -15,16 +15,16 @@ class ReminderStore extends Store<ReminderState> {
     const today = new Date();
     
     const initialSections: ReminderSectionData[] = [
-      { id: CATEGORIES.EVERYDAY, title: 'Everyday', isFixed: true, items: [
+      { id: Category.EVERYDAY, title: 'Everyday', isFixed: true, items: [
         { id: 1, text: '약 먹기', time: new Date(new Date(today).setHours(14, 0, 0, 0)), isAllDay: false, notified: true, done: true },
         { id: 2, text: '알고리즘 문제 풀기', time: new Date(new Date(today).setHours(16, 0, 0, 0)), isAllDay: false, notified: false, done: false },
         { id: 3, text: '산책하기', time: new Date(new Date(today).setHours(23, 0, 0, 0)), isAllDay: false, notified: false, done: true },
       ]},
-      { id: CATEGORIES.TODO, title: 'To Do', isFixed: true, items: [
+      { id: Category.TODO, title: 'To Do', isFixed: true, items: [
         { id: 4, text: '책 반납하기', isAllDay: true, notified: false, done: false },
         { id: 5, text: '편의점 택배 보내고 오기', time: new Date(new Date(today).setHours(16, 0, 0, 0)), isAllDay: false, notified: false, done: false },
       ]},
-      { id: CATEGORIES.WORK, title: 'Work', isFixed: false, items: [] },
+      { id: Category.WORK, title: 'Work', isFixed: false, items: [] },
     ];
 
     super({ sections: initialSections }, STORAGE_KEYS.REMINDER);
