--- src/shared/constants/category.ts
@@ -1,11 +1,0 @@
-export enum Category {
-  EVERYDAY = 'Everyday',
-  TODO = 'To Do',
-  WORK = 'Work',
-}
-
-export const CATEGORIES = [
-  { title: 'Everyday', value: Category.EVERYDAY },
-  { title: 'To Do', value: Category.TODO },
-  { title: 'Work', value: Category.WORK },
-];
